Use chrome.usb.getDevices instead findDevices
The former just lists devices, while the latter also opens them and (on
ChromeOS only) requests access. For some reason, the latter is failing
on my Debian system (and I suspect it might be trying the request access
thinking I'm running ChromeOS?).

Since the actual talking to the USB devices happens through the serial
API anyway, there is no reason to even try to open them, this code is
just intended to detect if the Pinoccio Scout is present in the first
place.