Improve handling of incorrect start-of-command bytes
Every command and reply in stk500 starts with 0x1b. When another byte
was received, before the serial read was aborted and an error message
was shown. However, since cbStep() doesn't do string formatting, the
variable part of the message (the actual byte received) was hidden.

This fixes that error message and continues to look for a 0x1b on an
error, which might make things slightly more reliable.